# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinVerify OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: List bank transactions
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankTransaction"
  /transactions/{id}/status:
    patch:
      operationId: updateTransactionStatus
      tags: [transactions]
      summary: Update transaction status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionStatusUpdate"
      responses:
        "200":
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankTransaction"
...
components:
  schemas:
...
    BankTransaction:
      type: object
      required: [id, date, narration, amount, type, source, status, confidenceScore]
      properties:
        id: { type: integer }
        date: { type: string }
        narration: { type: string }
        amount: { type: number }
        type: { type: string }
        source: { type: string }
        bankName: { type: ["string", "null"] }
        reference: { type: ["string", "null"] }
        status: { type: string }
        confidenceScore: { type: integer }
        matchedInvoiceId: { type: ["integer", "null"] }
        note: { type: ["string", "null"] }
    TransactionStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
        note: { type: ["string", "null"] }
...